+++ public/js/pie-chart.js
@@ -1,3 +1,16 @@
+const mqttHandler = require('../../mqtt_handler');
+
+let mqttClient = new mqttHandler();
+    mqttClient.connect();
+
+/** Get data from sensors */
+let getData = mqttClient.data();
+let rightflap = getData.rightflap;
+let leftflap = getData.leftflap;
+let rightshoulder = getData.rightshoulder;
+let leftshoulder = getData.leftshoulder;
+
+
 // Grab the bar chart canva
 const pieChartContext = document
   .getElementById('js-pie-chart')
@@ -9,15 +22,15 @@ const newPieChart = new Chart(pieChartContext, {
   type: 'pie',
   data: {
     // Chart Label Vertical
-    labels: ['Placeholder1', 'Placeholder2', 'Placeholder3', 'Placeholder4', 'Placeholder5'],
+    labels: ['Rightflap', 'Leftflap', 'Rightshoulder', 'Reftshoulder'],
     
-    // Colors to use #C472B9', '#E4CEE0', '#4382BB', '#84A6D6', '#DDF2F4'
+    // Colors to use #C472B9', '#E4CEE0', '#4382BB', '#84A6D6'
     datasets: [
       {
         // Chart data
-        data: [30, 24, 58, 64.1, 10.2],
+        data: [rightflap, leftflap, rightshoulder, leftshoulder],
         label: 'Pelvic Chair',
-        backgroundColor: ['#C472B9', '#E4CEE0', '#4382BB', '#84A6D6', '#DDF2F4'],
+        backgroundColor: ['#C472B9', '#E4CEE0', '#4382BB', '#84A6D6'],
         borderWidth: 0.3,
         borderColor: 'black',
       },
