+++ views/chart.ejs
